moved template parsing
